Revert "Remove non-functional hardware keyboard options."
The removed options are actually functional on QWERTY-phones and allow
users to disable auto-correction, auto-capitalization and auto-
punctuation when the hw-keyboard is used. So get these settings back
and change the CheckBox to a Switch while doing that. Also add
back the search-indexes for the options.

This reverts commits 572f90bdf8183e42c522fdf041b20b064ffc8549 and
bf55863b41efc3ad29c752f133f8739aac35d103.

Change-Id: I4cc37d1c0cde638b50a2ed532c37d031b5ed73d2
Signed-off-by: gamerman123x <19superguy98@gmail.com>