apps2sd: Enable apps2sd for apps with unspecified install location
Previously this would only be allowed if the user selected a
preferred install location requesting all apps possible be
installed on external storage.

Relax this by also allowing the user to choose to move an app
if they have the default value for the preferred install location
(default being to let the system pick).

Change-Id: Idf1091d12bd1f19d798a3e8522650de0115dae01
Signed-off-by: gamerman123x <19superguy98@gmail.com>