Settings: Safeguard against NPEs in lock screens
This commit makes the ConfirmLockPasswordFragment check whether there
is an activity to work with when handling actions. The NPEs there can
be triggered by triggering by making a call occur twice in rapid
succession so that the first call goes through fine and finishes the
activity with a result, but the second just throws as the activity
will have been finished already at that point and the fragment will
no longer be associated with it.

Change-Id: I0703c7b16694bbb0ad8df22908a076298245a900
Signed-off-by: gamerman123x <19superguy98@gmail.com>